// src/lib/schema.ts
export const ExibitionFormSchema = z.object({
  name: z.string().min(2, {
    message: "Name must be at least 2 characters."
  }),
  email: z.string().email({
    message: "Please enter a valid email address."
  }),
  phone: z.string().min(10, {
    message: "Please enter a valid phone number."
  }),
  company: z.string().min(2, {
    message: "Company name must be at least 2 characters."
  }),
  type: z.enum(["exhibitor", "visitor"])
});